requirements.txt
contourpy==1.3.3
cycler==0.12.1
dpcpp-cpp-rt==2025.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
ImageHash==4.3.2
intel-cmplr-lib-rt==2025.3.1
intel-cmplr-lib-ur==2025.3.1
intel-cmplr-lic-rt==2025.3.1
intel-opencl-rt==2025.3.1
intel-openmp==2025.3.1
intel-pti==0.15.0
intel-sycl-rt==2025.3.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mkl==2025.3.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
onemkl-license==2025.3.0
onemkl-sycl-blas==2025.3.0
onemkl-sycl-dft==2025.3.0
onemkl-sycl-lapack==2025.3.0
onemkl-sycl-rng==2025.3.0
onemkl-sycl-sparse==2025.3.0
opencv-python==4.13.0.90
pillow==12.0.0
polars==1.37.1
polars-runtime-32==1.37.1
pyelftools==0.32
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
scipy==1.17.0
setuptools==80.10.1
six==1.17.0
sympy==1.14.0
tbb==2022.3.0
tcmlib==1.4.1
torch==2.10.0+xpu
torchaudio==2.10.0+xpu
torchvision==0.25.0+xpu
triton-xpu==3.6.0
ultralytics==8.4.10
ultralytics-thop==2.0.18
umf==1.0.2
wheel==0.46.3
